% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGalaxyTool.R
\name{MakeGalaxyTool}
\alias{MakeGalaxyTool}
\title{Create a galaxy tool from a function in this package}
\usage{
MakeGalaxyTool(funcName, galaxyHome = "inst/galaxy", toolDir = "mytool")
}
\arguments{
\item{galaxyHome}{(OPTIONAL) Galaxy home folder, don't change unless you know what you're doing.}

\item{toolDir}{(OPTIONAL) tool directory relative to 'galaxyHome/tools/', use this to group tools within 1 folder.}

\item{func}{Function to generate a script of. Needs to be exported by this package.}
}
\value{
Output from the functional test of \code{func}
}
\description{
Create a galaxy tool consisting of an XML file and an R script.
The R script should be a function defintion of a documented function, using 
\pkg{roxygen2}. The documentation is then used to make the tool definition.
This function will also attempt to create a functional test, although this 
is likely to fail the first time since you need to add the data to the right
directory.
}
\examples{
\dontrun{
# Set your working directory to the root of this package
#   Make sure you have cloned this package to run this example.
MakeGalaxyTool(CreateDESeqData)
}
}
