\name{runFunctionalTest}
\alias{runFunctionalTest}
\title{Run the functional test associated with a function.}
\usage{
  runFunctionalTest(func, functionalTestDirectory)
}
\arguments{
  \item{func}{A function to be exposed in Galaxy.}

  \item{functionalTestDirectory}{Where to find functional
  tests, if func is not in a package.}
}
\value{
  Whether the test passes.
}
\description{
  FIXME
}

